Consolidate integration test setup into a dedicated crate
The code required to setup a complete test environment is rather a lot.
What started out as a mildly-unpleasant amount of copy/pasting between
the integration tests and the benchmarks has become very unpleasant.
What is more, there will be at least one more crate for the PKCS #11
module and that will also need to set up a complete test environment.

Thus, break the test setup code out into a siguldry-test crate. This
crate is marked as "publish = false" and users can add it to their
dev-dependencies by path.

This does mean running the integration tests doesn't work from the
published crates, but I think that's probably acceptable for now.
