+++ siguldry/Cargo.toml
@@ -129,6 +129,10 @@ assert_cmd = "2"
 criterion = "0.8"
 proptest = "1.6"
 
+[dev-dependencies.siguldry-test]
+version = "0.1"
+path = "../siguldry-test"
+
 [dev-dependencies.tracing-test]
 version = "0.2"
 # env filtering breaks usage in integration tests
